# backend/swagger/swagger.yaml
openapi: 3.1.0
info:
  title: ASAP HVAC Sample API
  version: 0.2.0
  description: |
    Sample REST API for the ASAP HVAC web demo. Provides CRUD support for
    HVAC operational data with pagination, filtering, and validation via Zod
    schemas. Designed to be upgraded to relational storage in future iterations.
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
  /api/customers:
    get:
      summary: List customers
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: planTier
          in: query
          description: Filter by assigned maintenance plan tier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: email
          in: query
          description: Filter by customer email address (case-insensitive match).
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: A paginated collection of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomer'
    post:
      summary: Create a new customer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerInput'
      responses:
        '201':
          description: Newly created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/customers/{id}:
    parameters:
      - $ref: '#/components/parameters/CustomerId'
    get:
      summary: Retrieve a customer
      tags: [Customers]
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update customer (idempotent)
      description: Performs an idempotent update accepting any subset of customer fields.
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a customer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a customer
      tags: [Customers]
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/appointments:
    get:
      summary: List appointments
      tags: [Appointments]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          description: Filter by appointment status.
          schema:
            type: array
            items:
              type: string
              enum: [scheduled, completed, canceled]
          style: form
          explode: true
        - name: technician
          in: query
          description: Filter by technician name.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: quoteRequestId
          in: query
          description: Filter by quote request identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAppointment'
    post:
      summary: Create an appointment
      tags: [Appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentInput'
      responses:
        '201':
          description: Newly created appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/appointments/{id}:
    parameters:
      - $ref: '#/components/parameters/AppointmentId'
    get:
      summary: Retrieve an appointment
      tags: [Appointments]
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update an appointment
      tags: [Appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update an appointment
      tags: [Appointments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdate'
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an appointment
      tags: [Appointments]
      responses:
        '204':
          description: Appointment deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /api/inspections:
    get:
      summary: List inspections
      tags: [Inspections]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: quoteRequestId
          in: query
          description: Filter by quote request identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: technician
          in: query
          description: Filter by technician name.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of inspections
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInspection'
    post:
      summary: Create an inspection
      tags: [Inspections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionInput'
      responses:
        '201':
          description: Newly created inspection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/inspections/{id}:
    parameters:
      - $ref: '#/components/parameters/InspectionId'
    get:
      summary: Retrieve an inspection
      tags: [Inspections]
      responses:
        '200':
          description: Inspection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update an inspection
      tags: [Inspections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionUpdate'
      responses:
        '200':
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update an inspection
      tags: [Inspections]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionUpdate'
      responses:
        '200':
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an inspection
      tags: [Inspections]
      responses:
        '204':
          description: Inspection deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /api/final-quotes:
    get:
      summary: List final quotes
      tags: [Final quotes]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: quoteRequestId
          in: query
          description: Filter by quote request identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: status
          in: query
          description: Filter by quote decision status.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/FinalQuoteStatus'
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of final quotes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFinalQuote'
    post:
      summary: Create a final quote
      tags: [Final quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalQuoteInput'
      responses:
        '201':
          description: Newly created final quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalQuote'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/final-quotes/{id}:
    parameters:
      - $ref: '#/components/parameters/FinalQuoteId'
    get:
      summary: Retrieve a final quote
      tags: [Final quotes]
      responses:
        '200':
          description: Final quote details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalQuote'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a final quote
      tags: [Final quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalQuoteUpdate'
      responses:
        '200':
          description: Updated final quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalQuote'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a final quote
      tags: [Final quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalQuoteUpdate'
      responses:
        '200':
          description: Updated final quote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalQuote'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a final quote
      tags: [Final quotes]
      responses:
        '204':
          description: Final quote deleted successfully
        '400':
          description: Final quote cannot be deleted while related invoices or parts orders exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based indexing).
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Number of records per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: Sort expression in the format `<field>[:asc|desc]`.
      schema:
        type: string
    CustomerId:
      name: id
      in: path
      required: true
      description: Unique identifier for the customer
      schema:
        type: string
...
    AppointmentId:
      name: id
      in: path
      required: true
      description: Unique identifier for the appointment
      schema:
        type: string
    InspectionId:
      name: id
      in: path
      required: true
      description: Unique identifier for the inspection
      schema:
        type: string
    FinalQuoteId:
      name: id
      in: path
      required: true
      description: Unique identifier for the final quote
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Customer email must be unique.
        details:
          description: Optional error metadata
          nullable: true
    ListMeta:
      type: object
      required: [total, page, limit, hasNextPage]
      properties:
        total:
          type: integer
          example: 45
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        hasNextPage:
          type: boolean
          example: true
    PaginatedCustomer:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        meta:
          $ref: '#/components/schemas/ListMeta'
    Customer:
      type: object
      required:
        - id
        - name
        - primaryAddress
        - email
        - phone
      properties:
        id:
          type: string
          example: cust-001
        name:
          type: string
          example: Jane Smith
        primaryAddress:
          type: string
          example: 123 Houston Ave, Houston, TX
        email:
          type: string
          format: email
          example: jane@example.com
        phone:
          type: string
          example: 281-555-0110
        planTier:
          type: string
          nullable: true
          example: Premium
    CustomerInput:
      type: object
      required:
        - name
        - primaryAddress
        - email
        - phone
      properties:
        name:
          type: string
        primaryAddress:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        planTier:
          type: string
          nullable: true
    CustomerUpdate:
      type: object
      description: At least one property is required when updating an existing customer.
      properties:
        name:
          type: string
        primaryAddress:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        planTier:
          type: string
          nullable: true
...
    PaginatedAppointment:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        meta:
          $ref: '#/components/schemas/ListMeta'
    Appointment:
      type: object
      required: [id, quoteRequestId, scheduledDate, window, technician, status]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        scheduledDate:
          type: string
          format: date
        window:
          type: string
        technician:
          type: string
        status:
          type: string
          enum: [scheduled, completed, canceled]
    AppointmentInput:
      type: object
      required: [quoteRequestId, scheduledDate, window, technician, status]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        scheduledDate:
          type: string
          format: date
        window:
          type: string
        technician:
          type: string
        status:
          type: string
          enum: [scheduled, completed, canceled]
    AppointmentUpdate:
      allOf:
        - $ref: '#/components/schemas/AppointmentInput'
    PaginatedInspection:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        meta:
          $ref: '#/components/schemas/ListMeta'
    Inspection:
      type: object
      required: [id, quoteRequestId, technician, findings, adjustments, recommendedServices]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        technician:
          type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/InspectionFinding'
        adjustments:
          type: array
          items:
            $ref: '#/components/schemas/InspectionAdjustment'
        recommendedServices:
          type: array
          items:
            type: string
    InspectionFinding:
      type: object
      required: [code, description, severity]
      properties:
        code:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, moderate, high]
    InspectionAdjustment:
      type: object
      required: [description, cost]
      properties:
        description:
          type: string
        cost:
          type: number
    InspectionInput:
      type: object
      required: [quoteRequestId, technician, findings]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        technician:
          type: string
        findings:
          type: array
          items:
            $ref: '#/components/schemas/InspectionFinding'
        adjustments:
          type: array
          items:
            $ref: '#/components/schemas/InspectionAdjustment'
        recommendedServices:
          type: array
          items:
            type: string
    InspectionUpdate:
      allOf:
        - $ref: '#/components/schemas/InspectionInput'
    FinalQuoteStatus:
      type: string
      enum: [draft, awaiting-approval, approved, expired, declined]
    PaginatedFinalQuote:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FinalQuote'
        meta:
          $ref: '#/components/schemas/ListMeta'
    FinalQuote:
      type: object
      required: [id, quoteRequestId, baseEstimate, adjustmentsTotal, finalTotal, expiresOn, status]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        baseEstimate:
          type: number
        adjustmentsTotal:
          type: number
        finalTotal:
          type: number
        expiresOn:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/FinalQuoteStatus'
    FinalQuoteInput:
      type: object
      required: [quoteRequestId, baseEstimate, adjustmentsTotal, finalTotal, expiresOn, status]
      properties:
        id:
          type: string
        quoteRequestId:
          type: string
        baseEstimate:
          type: number
        adjustmentsTotal:
          type: number
        finalTotal:
          type: number
        expiresOn:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/FinalQuoteStatus'
    FinalQuoteUpdate:
      allOf:
        - $ref: '#/components/schemas/FinalQuoteInput'
...